+++ changelog.md
@@ -1,5 +1,6 @@
 
 TODO:
+	token organization issue: voice result new row
 	symbols of NUMBER & TIME_SIG
 	staff svg parser debug page
 	import user MIDI for sheet